Async Redis has to be setup for each test
There are a few requirements here:
1. Most important is that we flush redis before any of the tests are
executed. This can only be done using a normal synchronous call
since in the setUpClass() method, the event loop is not running yet.

2. Next in importance is that the async redis object MUST be instantiated
in the event loop that it will work with. I cannot span multiple event loops
by having a long life. It must be short and quick. So that is what I have done
here by moving the call to instantiate the async redis client to an async
function. Previously it was instantiated in a method that was not async and
that did not have an event loop associated with it.

Now these unit tests run perfectly fine.